move register method to users controller
